Updating provider to support rancher 2.5

+++ scripts/gotestacc_docker.sh
@@ -10,9 +10,11 @@ DOCKER_URL="https://download.docker.com/linux/static/stable/x86_64/docker-17.03.
 DOCKER_BIN=$(which ${DOCKER_NAME} || echo none)
 if [ "${DOCKER_BIN}" == "none" ] ; then
   export DOCKER_BIN=${TESTACC_TEMP_DIR}/${DOCKER_NAME}
-  curl -sL ${DOCKER_URL} | tar -xzf - 
-  mv docker/docker ${DOCKER_BIN} && rm -rf docker
-  chmod 755 ${DOCKER_BIN}
+  if [ ! -x "${DOCKER_BIN}" ]; then
+    curl -sL ${DOCKER_URL} | tar -xzf - 
+    mv docker/docker ${DOCKER_BIN} && rm -rf docker
+    chmod 755 ${DOCKER_BIN}
+  fi
 fi
 
 BUILDER_TAG=${BUILDER_TAG:-"terraform-provider-rancher2_builder"}
