+++ scripts/gotestacc_vars.sh
@@ -22,9 +22,11 @@ DOCKER_URL="https://download.docker.com/linux/static/stable/x86_64/docker-17.03.
 DOCKER_BIN=$(which ${DOCKER_NAME} || echo none)
 if [ "${DOCKER_BIN}" == "none" ] ; then
   export DOCKER_BIN=${TEMP_DIR}/${DOCKER_NAME}
-  curl -sL ${DOCKER_URL} | tar -xzf - 
-  mv docker/docker ${DOCKER_BIN} && rm -rf docker
-  chmod 755 ${DOCKER_BIN}
+  if [ ! -x "${TEMP_DIR}/${DOCKER_NAME}" ]; then
+    curl -sL ${DOCKER_URL} | tar -xzf - 
+    mv docker/docker ${DOCKER_BIN} && rm -rf docker
+    chmod 755 ${DOCKER_BIN}
+  fi
 fi
 DOCKER_LIST_NAME=${DOCKER_LIST_NAME:-"docker_ids"}
 DOCKER_LIST=${TEMP_DIR}"/"${DOCKER_LIST_NAME}
